requirements.txt
llama_cpp_python==0.2.82
huggingface-hub==0.23.4
gradio==4.38.1